# app/Models/Passenger.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

/**
 * @OA\Schema(
 *   schema="Passenger",
 *   type="object",
 *   @OA\Property(property="id", type="integer"),
 *   @OA\Property(property="name", type="string"),
 *   @OA\Property(property="email", type="string")
 * )
 */
class Passenger extends Model
{
    protected $fillable = [
        'name',
        'document'
    ];
}