# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkyLink API",
        "description": "API para gestionar reservas y pasajeros con actualizaciones en tiempo real",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Listar reservas",
                "operationId": "f238872d1695b5212f78f33c5abf654a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (PENDING, CONFIRMED, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva",
                "operationId": "5841b87b2610c0159482fd94e48289ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "flight_number",
                                    "departure_time",
                                    "passengers"
                                ],
                                "properties": {
                                    "flight_number": {
                                        "type": "string",
                                        "example": "SKY123"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-09-12T18:00:00Z"
                                    },
                                    "passengers": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "document"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Juan Perez"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "12345678"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Passenger": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "flight_number": {
                        "type": "string"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "CONFIRMED",
                            "CANCELLED",
                            "CHECKED_IN"
                        ]
                    },
                    "passengers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Passenger"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Reservations",
            "description": "Reservations"
        }
    ]
}